CI fixes

+++ src/internal/args.c
@@ -25,7 +25,7 @@ static void
 FieldTypePoolGBToBytesSet(void *field, const YAMLnode *node)
 {
    double gb          = strtod(node->mapped_val, NULL);
-   *((double *)field) = gb * GB_TO_BYTES;
+   *((double *)field) = gb * (double)GB_TO_BYTES;
 }
 
 #define General_FIELDS(_prefix)                                                      \
@@ -97,10 +97,10 @@ hypredrv_GeneralSetDefaultArgs(General_args *args)
    args->use_vendor_spmv   = 0;
 #endif
    args->num_repetitions  = 1;
-   args->dev_pool_size    = 2.0 * GB_TO_BYTES;
-   args->uvm_pool_size    = 2.0 * GB_TO_BYTES;
-   args->host_pool_size   = 2.0 * GB_TO_BYTES;
-   args->pinned_pool_size = 0.1 * GB_TO_BYTES;
+   args->dev_pool_size    = 2.0 * (double)GB_TO_BYTES;
+   args->uvm_pool_size    = 2.0 * (double)GB_TO_BYTES;
+   args->host_pool_size   = 2.0 * (double)GB_TO_BYTES;
+   args->pinned_pool_size = 0.1 * (double)GB_TO_BYTES;
 }
 
 void
